refactor: move eval expression test
This test should be at the start of the blocking transition tests
because it's only the first part. Creating the relevant job.
